F/147429 medicos

+++ src/main/java/es/plexus/hopes/hopesback/repository/model/Role.java
@@ -16,16 +16,16 @@ import javax.persistence.Table;
 @Table(name = "roles")
 public class Role {
 
-    @Id
-    @Column(name = "rol_id")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+	@Id
+	@Column(name = "rol_id")
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
 
-    @Enumerated(EnumType.STRING)
-    @Column(name="rol_name", length = 50)
-    private ERole name;
+	@Enumerated(EnumType.STRING)
+	@Column(name = "rol_name", length = 50)
+	private ERole name;
 
-    @Column(name="rol_description", length = 500)
-    private String description;
+	@Column(name = "rol_description", length = 500)
+	private String description;
 
 }
